# Conflicts:
#	app/gtnForecasting/src/main/java/com/stpl/app/gtnforecasting/salesprojection/utils/HeaderUtils.java

+++ app/gtnAdminConsole/src/main/java/com/stpl/app/adminconsole/processscheduler/logic/tableLogic/HierarchyTableLogic.java
@@ -40,7 +40,7 @@ public class HierarchyTableLogic extends PageTableLogic {
         int count = 0;
         try {
             if (isFirstLoad) {
-                count = (Integer) searchLogic.getHierarchyDefinitionCount(binder, this.getFilters(), hierarchyType);
+                count = searchLogic.getHierarchyDefinitionCount(binder, this.getFilters(), hierarchyType);
             }
             isResultsEmpty = count == 0;
             count = isReset ? 0 : count;
